meta: Split device drivers in their own files
Before, there was one big file per vendor for all their devices, now
each device has its own file.

Considering that with only two devices in a single file, it started to
become unmanagable... I think this is the right move!

This also updates the build script because for some reason, that change
broke it :/