drivers/*: Reworked action system
Now there is a single create_actions() function for each driver which
handles creating the action and the handlers, avoiding the massive
if-else block there was before in the run_action() function.

I also editted some of the name, and re-did some of the style for
everything to be consistent. I also removed the action for color
changing on the Rival 3 Wireless, as this was mostly an experiment and
never worked properly.

Next, I'll work on adding macros and utils to make it even easier and
shorter to write actions. Another cool thing would be to have some
automatic parsing of the arguements, though I have no idea how to do
that yet :/